+++ alicloud/resource_alicloud_instance_test.go
@@ -1572,7 +1572,7 @@ func testAccCheckInstanceImageOrigin(common string) string {
 		name_regex = "^centos_6\\w{1,5}[64]{1}.*"
 	}
 	variable "name" {
-		default = "testAccCheckInstanceImageOrigin"
+		default = "tf-testAccCheckInstanceImageOrigin"
 	}
 
 	resource "alicloud_instance" "update_image" {